Link imported ObjC forward declarations to their Swift definition
On occasion, an Objective-C header will forward declare an interface
or protocol that is defined elsewhere in Swift. If the definition were
in a Clang module instead, the forward declaration would be linked to
its definition. Prior to this patch, this was not the case for @objc
exposed Swift declarations. This patch fixes that.