[move-only] Change MoveOnlyBorrowToDestructureTransform into a driver invoked by the tester/main checker pass rather than a utility.
This is in preparation for sinking the main computation of the pass into an
implementation local helper struct that can process multiple values. I am doing
this so we can use the same code to process switch_enum arguments and am using
the current implementation to make sure I don't break anything.