Add a test to verify that AnyHashable gets eagerly unwrapped
To allow for better compiler optimizations, the
runtime should guarantee that AnyHashable does
not get boxed unnecessarily.

In particular, there was a concern that casting
an `AnyHashable` containing a class reference
to `AnyObject` would box the AnyHashable instead
of unwrapping it.

Further testing shows that the runtime is NOT
doing this unnecessary boxing at current.

Resolves rdar://90269352