Guard the lifetime of StringObject.nativeStorage in mutating methods
StringGuts has mutating methods that access StringObject.nativeStorage
while modifying self as such:

  self = _StringGuts(self._object.nativeStorage)

`self._object` can be destroyed immediately after its last use. We
can't rely on `self` having a lexical lifetime within a mutating
method. This follows from the fact that `self` is effectively `inout`
within this scope, and thus conceptually destroyed on function entry.

Since the nativeStorage property generates a new reference out of thin
air (from an integer bitcast), `object` may be destroyed before the new
reference is retained.

Fix this by creating a separate copy of object with its own lexical
scope that covers the use of `nativeStorage`:

  var object = self._object
  self = _StringGuts(object.nativeStorage)

Additionally add an explicit `withExtendedLifetime` scope. This forces
the compiler to keep 'object' alive even when the standard library is
not compiled with "lexical lifetimes" enabled, or if we choose to
optimize _StringObject using eager-consume semantics. We may decide to
remove the withExtendedLifetime in the future.