IRGen: correct Twine handling
Twine is an intermediate reference and should not be constructed to hold the value.  An unowned string piece should be referenced by `StringRef`.  This now provides storage for the name.  We would previously see invalid memory usage due to the Twine being destroyed before the use in the `CreateLoad` invocation.