[IDE] Check whether an AST node contains the IDE inspection point based on `CharSourceRange`
This fixes an issue if the range ends with a string literal that contains the IDE inspection target. In that case the end of the range will point to the start of the string literal but the IDE inspection target is inside the string literal and thus after the range’s end.