[Concurrency] Memory order of success must be >= failure path in exchange
This triggers a C++ stdlib assertion and must be fixed.
We don't rely on the status as a barrier here, because we didn't get to
a fully lockless task group yet and the group is protected with a lock.
Either way, changing from relaxed to release here would be stronger and
not weaker, so it is safe to do either way.