Serialization: avoid pushing static libraries into autolink
Remove the static libraries from autolinking.  This should be safe to do
with SPM builds which does not respect the static linking and instead
treats all targets as object libraries for static linking.  With CMake
this should also not be a problem as it already requires dependencies be
spelt out and then correctly forms the linker invocation.  This prevents
us from double linking a library (both into a potential intermediate
shared library and any consumers of any public interfaces).