[Test] Replaced argument with call result.
The test verifies the behavior of CanonicalizeOSSALifetime on adjacent
phis.  Previously, it was being run on a phi one of whose incoming
values was lexical.  Now that phis any of whose transitive incoming
values are lexical are themselves recognized as lexical, the lifetime is
not being canonicalized.  When lexical lifetimes are canonicalized, this
will not be necessary.